refactor: Change directory location of User entity
